should fix some undefined behavior

+++ cocos/2d/CCNode.cpp
@@ -274,7 +274,7 @@ void Node::setLocalZOrder(int z)
 /// used internally to alter the zOrder variable. DON'T call this method manually
 void Node::_setLocalZOrder(int z)
 {
-    _localZOrderAndArrival = (static_cast<std::int64_t>(z) << 32) | (_localZOrderAndArrival & 0xffffffff);
+    _localZOrderAndArrival = static_cast<std::int64_t>(static_cast<std::uint64_t>(z) << 32) | (_localZOrderAndArrival & 0xffffffff);
     _localZOrder = z;
 }
